requirements.txt
fastapi
uvicorn
torch
transformers
scikit-learn
numpy
bs4
python-magic
pydantic

